+++ R/visualization.R
@@ -559,6 +559,7 @@ RidgePlot <- function(
 #' single violin shapes.
 #' @param adjust Adjust parameter for geom_violin
 #' @param flip flip plot orientation (identities on x-axis)
+#' @param add.noise determine if adding a small noise for plotting
 #' @param raster Convert points to raster format. Requires 'ggrastr' to be installed.
 # default is \code{NULL} which automatically rasterizes if ggrastr is installed and
 # number of points exceed 100,000.
@@ -597,6 +598,7 @@ VlnPlot <- function(
   combine = TRUE,
   fill.by = 'feature',
   flip = FALSE,
+  add.noise = TRUE,
   raster = NULL
 ) {
   if (
@@ -633,6 +635,7 @@ VlnPlot <- function(
     combine = combine,
     fill.by = fill.by,
     flip = flip,
+    add.noise = add.noise,
     raster = raster
   ))
 }
@@ -5560,6 +5563,7 @@ Col2Hex <- function(...) {
 # ggplot object. If \code{FALSE}, return a list of ggplot objects
 # @param fill.by Color violins/ridges based on either 'feature' or 'ident'
 # @param flip flip plot orientation (identities on x-axis)
+# @param add.noise determine if adding a small noise for plotting
 # @param raster Convert points to raster format, default is \code{NULL} which
 # automatically rasterizes if plotting more than 100,000 cells
 #
@@ -5591,6 +5595,7 @@ ExIPlot <- function(
   combine = TRUE,
   fill.by = NULL,
   flip = FALSE,
+  add.noise = TRUE,
   raster = NULL
 ) {
   assay <- assay %||% DefaultAssay(object = object)
@@ -5676,6 +5681,7 @@ ExIPlot <- function(
       pt.size = pt.size,
       log = log,
       fill.by = fill.by,
+      add.noise = add.noise,
       flip = flip
     ))
   }
@@ -5693,6 +5699,7 @@ ExIPlot <- function(
         cols = cols,
         pt.size = pt.size,
         log = log,
+        add.noise = add.noise,
         raster = raster
       ))
     }
@@ -6384,6 +6391,7 @@ MultiExIPlot <- function(
   seed.use = 42,
   log = FALSE,
   fill.by = NULL,
+  add.noise = TRUE,
   flip = NULL
 ) {
   if (!(fill.by %in% c("feature", "ident"))) {
@@ -6450,6 +6458,9 @@ MultiExIPlot <- function(
     data$expression <- data$expression + 1
   } else {
     noise <- rnorm(n = nrow(x = data)) / 100000
+  }
+  if (!add.noise) {
+    noise <- noise*0
   }
   for (f in unique(x = data$feature)) {
     if (all(data$expression[(data$feature == f)] == data$expression[(data$feature == f)][1])) {
@@ -7213,6 +7224,7 @@ SingleDimPlot <- function(
 #' @param cols Colors to use for plotting
 #' @param seed.use Random seed to use. If NULL, don't set a seed
 #' @param log plot Y axis on log10 scale
+#' @param add.noise determine if adding small noise for plotting
 #' @param raster Convert points to raster format. Requires 'ggrastr' to be installed.
 #' default is \code{NULL} which automatically rasterizes if ggrastr is installed and
 #' number of points exceed 100,000.
@@ -7242,6 +7254,7 @@ SingleExIPlot <- function(
   cols = NULL,
   seed.use = 42,
   log = FALSE,
+  add.noise = TRUE,
   raster = NULL
 ) {
    if (!is.null(x = raster) && isTRUE(x = raster)){
@@ -7285,6 +7298,9 @@ SingleExIPlot <- function(
     data[, feature] <- data[, feature] + 1
   } else {
     noise <- rnorm(n = length(x = data[, feature])) / 100000
+  }
+  if (!add.noise) {
+    noise <-  noise * 0
   }
   if (all(data[, feature] == data[, feature][1])) {
     warning(paste0("All cells have the same value of ", feature, "."))